pyproject.toml
[project]
name = "scenari-sumo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "colorama>=0.4.6",
    "numpy>=2.3.3",
    "pyyaml>=6.0.3",
    "sumolib>=1.24.0",
    "traci>=1.24.0",
]
